+++ src/index.js
@@ -4,17 +4,17 @@ const config = require('./configuration');
 const compileSource = require('./compile-source');
 const { transform } = require('./transformations/transform');
 const packageJson = require('../package.json');
+const { calculateChecksum } = require('./utils');
 
 /**
  * Prepares list header
  *
  * @param {*} configuration - compilation configuration.
-See the repo README for the details on it.
+ * See the repo README for the details on it.
  * @returns {Array<string>} header lines
  */
 function prepareHeader(configuration) {
     const lines = [
-        '!',
         `! Title: ${configuration.name}`,
     ];
 
@@ -67,7 +67,7 @@ function prepareSourceHeader(source) {
  * Compiles a filter list using the specified configuration.
  *
  * @param {*} configuration - compilation configuration.
-See the repo README for the details on it.
+ * See the repo README for the details on it.
  * @returns {Promise<Array<string>>} the array of rules.
  */
 async function compile(configuration) {
@@ -101,8 +101,12 @@ async function compile(configuration) {
 
     // Now prepend the list header and we're good to go
     const header = prepareHeader(configuration);
-    consola.info(`Final length of the list is ${header.length + finalList.length}`);
-    return header.concat(finalList);
+    // Calculate checksum
+    const checksum = calculateChecksum(header, finalList);
+    // Concat everything together
+    const data = ['!', checksum, ...header, ...finalList];
+    consola.info(`Final length of the list is ${data.length}`);
+    return data;
 }
 
 module.exports = compile;